refactor: 내부 오류 응답 리팩토링
