+++ src/main/java/com/recipe/app/common/exception/BaseException.java
@@ -1,18 +1,14 @@
 package com.recipe.app.common.exception;
 
 import com.recipe.app.common.response.BaseResponseStatus;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.Setter;
 
 @Getter
-@Setter
-@AllArgsConstructor
 public class BaseException extends RuntimeException {
-    private BaseResponseStatus status;
-    private String ingredientName;
+    private final BaseResponseStatus status;
 
     public BaseException(BaseResponseStatus status) {
+        super(status.getMessage());
         this.status = status;
     }
 }
